+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fixed: Windows packages increased in size. ([#713](https://github.com/ksnip/ksnip/issues/713))
 * Fixed: Translations are not working in ksnip snap package. ([#662](https://github.com/ksnip/ksnip/issues/662))
 * Fixed: The string 'Actions' is not available for translation. ([#729](https://github.com/ksnip/ksnip/issues/729))
+* Fixed: HiDPI issue with multiple screen on Windows. ([#668](https://github.com/ksnip/ksnip/issues/668))
 * Fixed kImageAnnotator: Crash while typing text on wayland. ([#256](https://github.com/ksnip/kImageAnnotator/issues/256))
 
 ## Release 1.9.1